fix: moved files to separate directories
